feat: ProviderScope の設定
